src/external/handlers/filesources.ts
import { ExternalDataState } from '../../../spec/ExternalData';
import { TransformationError } from '../errors';
import { handleError } from './base';

export async function readFile(state: ExternalDataState, sourceId: string): Promise<any[]> {
  const source = state.sources.file.get(sourceId);
  if (!source) {
    throw new TransformationError(`File source ${sourceId} not found`);
  }

  try {
    // Placeholder for actual file reading
    // In a real implementation, this would:
    // 1. Read file based on type
    // 2. Parse content
    // 3. Apply transformations
    return [];
  } catch (error) {
    throw handleError(error, TransformationError, 'Failed to read file');
  }
}

export async function validateFile(state: ExternalDataState, sourceId: string): Promise<boolean> {
  const source = state.sources.file.get(sourceId);
  if (!source) {
    throw new TransformationError(`File source ${sourceId} not found`);
  }

  try {
    // Placeholder for actual file validation
    // In a real implementation, this would:
    // 1. Check file exists
    // 2. Validate format
    // 3. Check permissions
    return true;
  } catch (error) {
    throw handleError(error, TransformationError, 'Failed to validate file');
  }
}